Distribute the utils.h header file
The set of utilities contained in utils.h/cpp is a mix of things that
are used mostly internally within libprofit, and of others that can be
useful to third parties. The functions were all PROFIT_API-decorated
already, so it's little effort to actually make them part of an
installation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>