Adding lib initialization fixture to unit tests
Until now unit tests never initialized or finalized libprofit, which in
theory is a requirement for the correct functioning of the library. So
far this has had no visual impact on the tests because in practice
initializing libprofit makes a difference only when using FFT-based
convolution and OpenCL computations, and only to a certain degree.
However, with the valgrind-based test runs it's become clear that we
needed to be more correct on our unit testing if we wanted to have
better results out of valgrind.

This commit adds a new global fixture that initializes and finished
libprofit before each test suite is ran. In order to avoid changing any
existing ~/.profit directories we use a new, temporary directory that
gets deleted afterwards. We init/finish at the test suite level instead
of at individual test level simply to avoid the cost of hitting the
filesystem unnecessarily.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>