New internal way of setting parameter values
The previous implementation was unnecessarily wordy, with subclasses
having to implement too much logic which was extremely similar from case
to case.

With this new design we now simply register the private members of each
profile into the base class' parameter holder maps, which are then used
to set the values on the variables when a parameter value is given by
the user. This centralization will also allow us in the future to do
other fancy things like listing all supported parameters on a given
profile, or attach extra information to each parameter.

Also, instead of using the old <typeinfo> header we now use the new type
traits, which give us exactly the string we want for each type.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>