New unknown_parameter exception
This is a more specific type of invalid_parameter, which means that the
given parameter is unknown to the profile. This could have been used by
the current set of parameter() methods on the Profile class, but they
are already throwing invalid_parameter in those cases, and we probably
don't want to break that.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>