New parameter setting method on Profile
This new method works with a single string, which is decomposed into a
name=value pair and then lexically casted to the required type depending
on the parameter name. Internally a templated method (which uses the new
type traits for getting type names) is used to avoid code repetition.

Unlike the other parameter() methods, this one throws both
invalid_parameter *and* unknown_parameter so users can decide what to do
in both cases separately.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>