profit-cli more C++11-ish
The original version of the program was written way back when libprofit
was almost pure C, or when I was still learning my way through the std
library. This mix of C and C++ constructs was showing its age, specially
because is one of the things that prevents profit-cli from compiling
under Windows systems.

This commits encompasses the removal of most of the C-based code, and
replaces it with a corresponding C++11 version, which should take us
into the right direction to start compiling profit-cli with Windows.
The main exception are the fits-related methods, which are going into a
separate module and will be changed after that.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>