Fixing naming of OpenCL kernel cache entries
Firstly, we need to sanitize the platform and device names so they are
valid file names (i.e., don't contain slashes, colons, etc). Such
characters are being replaced now with underscores. On top of that we
are also turning spaces into underscores for uniformity. And on top of
that we are also separating the platform/device name and its supported
OpenCL version with an extra underscore.

This lack of filename cleanup was preventing some platforms to cache
their entries. For example, in my computer my AMD card's device, when
exposed via Mesa's Clover OpenCL platform, is named "AMD Radeon R7 M260
Series (OLAND / DRM 3.23.0 / 4.15.0-30-generic, LLVM 6.0.0)", which
prevented the code from caching it correctly (because of the "/"
characters).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>