Some documentation fixes/updates
The documentation generated by doxygen was most of the time not readable
by breathe because of the PROFIT_API macro, which we now expand to an
empty value. In the rst documentation itself some phrases were out of
date, and also many of the references to API entry points were giving
warnings due to us using the incorrect directive to link to their
definition.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>