OpenCL kernels stored by their own
Because we ship OpenCL source code as part of the libprofit library
(i.e., as part of the compiled shared library binary), we previously had
each of our .cl files containing not only the code for the OpenCL kernel
(or functions/definitions), but also the surrounding scafolding
necessary to wrap all this code as a C++ literal, #include it directly
into a C++ program and easily assign it into a variable. Although this
has been working fine, it is still not so nice that the sources are
cannot live entirely on their own.

With this commit I'm changing how things work. Now .cl files contain
*only* the OpenCL code, and for each of them a header file is generated
that declares a variable within the libprofit namespace containing the
whole of the OpenCL code. We can then safely include these files at the
beginning of the opencl.cpp module, and just refer to the variables
within the code.

In order to safely re-generate the header files when there are changes
in the corresponding .cl files I needed to use add_custom_command. On
the other hand I didn't want to rely on a sed or a bash script to do the
replacing work, so I wrote the generating logic as a cmake script under
the cmake/ subdirectory, and I'm invoking that script from the main
CMakeLists.txt.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>