Simplifying accuracy adjustment
When automatically adjusting the integration accuracy parameter in
radial profiles, most of them actually leave this untouched; therefore a
default implementation that simply returns the current accuracy actually
makes sense.

On the other hand, there is no need for subclasses to access the
this->acc member if we pass it as a parameter in the first place. Thus
we are changing the signature of adjust_acc to include the current
accuracy as a parameter.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>