Making profile members private
This doesn't affect things much, except in the case of the
RadialProfile, where we were exposing the members with protected for
convenice (so subclasses could address them directly). With the latest
changes most of these members are not needed anymore by subclasses, so
it's safe to move them to private. The exceptions are only a few
profiles (sersic, ferrer, moffat) which still need some of this
information. We are making them explicit friends of the RadialProfile
class, so in the general case we still don't have visibility.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>