Abstracting usage of std::cout and std::cerr
This would make the code more flexible in case we wanted to output
things into different places. Although this is a final goal indeed, in
reality we are only doing it to centralize the places where we reference
std::cout and std::cerr directly, therefore reducing the number of "code
smells" found by SonarQube.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>