profit-cli correctly handling unknown parameters
We used to handled them gracefully, only printing a warning message onto
the console but continuing with the process. When we introduced the new
set_parameter(const std::string &) we broke that though, and it wasn't
until now, that we are actually testing the profit-cli program
automatically, that we found out about this.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>