Re-implementing path_exists for Windows
I looked how boost-filesystem implements this, and they only check if
the file is a directory or not, and if it's not then it's assumed it is
a normal file (after checking for symlinks, which we don't support).
This is because FILE_ATTRIBUTE_NORMAL works only when it's the only
attribute present, which is probably the case most of the times?

This seems to be the cause for a profit-cli test failure in Windows,
specifically when trying to load a PSF from a fits file (which is not
recognized as a file because our code determined there was no file).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>