Fixing OpenCL_command_times::operator+
The operator+ method is based on the operator+= being applied to a new
object t1. This t1 object was not copy-initialized from *this though,
but default-initialized. This means that the operator+(other) member
function was not really adding "other" to "this" but to an empty value,
effectively returning an object with the same internal values as
"other".

Not a big deal, since we use these numbers only for inspecting
the performance of OpenCL operations; moreover, the operator+ function
was only used together with OpenCL filling events (which we don't use
much), so only those numbers were bogus.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>