Re-organizing ProfileStats classes and code
First of all, there is no need for user-written default constructors.
We use value initialization to create objects of these types (via
std::make_shared), and therefore their values will be zero-initialized
without us having to explicitly write support for that.

Secondly, the inner radial_subsampling_stats struct used in the
RadialProfileStats class is now declared separately.

Finally, a good comment about the need for the virtual destructor on the
ProfileStats class was necessary (because I thought it could have been
removed, but turned out to be wrong). I don't want to shoot myself on
the foot again on this one, so better to document it more clearly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>