Running sonar-qube analysis on Travis CI
This new integration uses the same coverage build we are using for
uploading results to coveralls. After uploading the results to coveralls
we now also analyze the code and upload the report to SonnarCloud.

For this to work we need to move the compilation step (i.e., "make all") from
the install to the script step. This is because the SonarCloud addon actually
runs just before before_script, so we can't rely on it being there at install
time. This is probably good anyway as compilation is part of what we "test" in
Travis.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>