Using new omp_2d_for function
This commit contains only the replacement of all macros/fors by the new
function call. Because now we don't have a double-for loop but only a
single lambda function body the level of indentation of the code needs
to be reduced by one, but that will go in the next commit to leave this
simple and clean.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>