Making gcc 4.6 happy
I'm not exactly sure why, but gcc 4.6 was previously happily compiling
the code as it was, but once we restructured it and broke it down into
individual functions it didn't understand the {} aggregate
initialization syntax anymore, even though the types had not changed in
the slightest.

Making gcc 4.6 happy makes ubuntu precise happy, which in turn makes
more users happy.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>