Lowering OpenMP version requirement
The only reason why we required OpenMP 3.0 was because we collapse two
"for" statements in a single "parallel for" pragma. It's easy enough to
imitate that with a single for, so if we find ourselves having only
OpenMP 2.0+ then we provide such single-for version of the code to
implement omp_2d_for.

Other than that, there is the extra requirement in OpenMP 2.0 that the
loop variable has to be signed, which we obey in our single-for version
of the omp_2d_for function template.

I don't know how much lower one can go in terms of OpenMP versions, but
this seems reasonable already, as it includes all major compilers: gcc,
clang, MSVC.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>