New dot product function implementations
Three new different and flexible implementations are shipped in now with
libprofit with the purpose of better using the CPUs that we have
available.

The first one is pure software, but it's templated to explore the
efficiency of packing together more sum/product operations depending on
the CPU being used. Experimental results (in my personal machine)
indicate that packing 4 operations yields the best result, and going up
to 8 degrades performance, the I'm leaving the implementation to use a
hard-coded value of 4. In the future we can tune this maybe depending on
the particular CPU being used.

The second and third implementations are based on x86-64 extensions
(SSE2 and AVX, respectively), so they are included only conditionally
after inspecting the system and checking whether the extensions are
supported or not. They both share the same templated code which
maintains vectorized variables all the way through the end, until the
final result needs to be extracted. Like in the case of the software
implementation we also template the code on the maximum number of
operations we want to pack together. Experimental results in my machine
show that using 2 SSE2 and 2 AVX registers (4 and 8 double values
respectively) yield the best results.

Both intrinsics-based implementations are faster than the software-only
implementation, and between the two intrinsics AVX is faster. This sets
the base as to what implementation is finally used by the "dot_product"
function, which is what the user should be using at the end of the day.

All code is static and inline, and declared in a header file, so there
should be no penalty for using this function within a tight loop as
there are no extra calls generated.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>