Removing unnecessary code
This was only necessary during our experiments, when we distinguished
between aligned and misaligned data access. We use unaligned loads
everywhere now so there's no point on keeping this code around. If in
the future we want to experiment a bit more with aligned access patterns
we can bring this back.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>