Simplifying main body of associative convolver
Among other things, I've removed unused variables, out-of-date comments,
and moved some 'if' statements out of the innermost loop, since they
don't depend on it. Thanks to this the way we first initialize, and then
advance the krn and src pointers within the innermost loop can be
greatly simplified to simpler expressions and constant shifts.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>