Moved, renamed and doc'ed intrinsic enumeration
The actual enumeration values to different CPU extended instruction sets
offering SIMD operations (so we're calling it "simd_instruction_set"
now) and not to the nature of the functions that give us access to those
operations (which are "intrinsic" functions, the old name of the
enumeration). We are not renaming the rest of the function names that
reference "instrinsic" because those actually make sense. We are also
not renaming the "Intrinsic" template parameters because otherwise this
commit would become way too verbose.

Moving the enumeration higher up and exposing all its
elements will give users the possibility to use this enumeration to
specify which SIMD instructions should be used by the brute-force
convolver. This is probably not useful in general (we still want users
to not have to care about these details), but in benchmarks and tests we
probably want to have this flexibility.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>