New instruction_set pref to create convolvers
This new preference is passed down to the associative brute-force
convolver (which is now templated). The template argument of the
convolver will then be carried out to the dot_product function (which
will also be templated), so the final dot product implementation can be
both selected at runtime, and be as efficient as possible. These
implementation exist already, but we don't have the flexibility to
choose between them at runtime.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>