Reducing dynamic memory allocations, part 1
By having a member Image variable at the FFTConvolver level we don't
need to generate a new extended Image object every time we extend the
source and kernel objects, but instead we can use the member objects to
receive the extended version of the images.

Because we are also using now the extended source image to receive the
result of the backward transform we need to zero it out before using it
again to receive the extended version of the next source image.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>