Fixed underlying type for enumeration
We previously had no explicit underlying type for our enumeration, and
therefore its range was defined by "the smallest bit field large enough
to hold all enumerators"; in this case, a 2-bit field. This therefore
might not allow us to reliably cast some integer values that were
outside the range established by the enumerator list into the
enumeration type because the result was unspecified (and UB since C++17
actually).

Having a fixed underlying type for the enumeration should allow this
type of casts work, and finally let us test reliably the bit of code
that we want to test (i.e., an effort_t enumerator that is not part of
the enumeration).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>