Improving memory management in FFTRealTransform
The class constructor had a bit of an issue. The situation was like
this: the unique_ptr objects were being empty initialized in the
constructor's initializer list, then in the constructor body buffers
were being allocated, and then given to the unique_ptr objects. Right
after that process takes place there is a call that can throw
(get_effort). When it does the unique_ptr objects are destroyed, as
part of the winding back of the constructor method. The issue was that
the assignment of the memory into the unique_ptr objects was not
required to have happened because the compiler was free to re-order some
of these statements. If that happens, then some of the memory that was
allocated but not wrapped into the unique_ptr would have been dangling.
I believe this was the cause of the error I was seeing in the MSVC
builds, and that could have affected other compilers anyway.

The solution is obviously to be more strict about how and when we
allocate and deallocate the memory wrapped by the unique pointers. With
this change we now: a) initialize the unique pointers in the initializer
list of the constructor, and b) use a deleter object for automatic
management of the proper destruction of the underlying pointer. This
second point makes it unnecessary also to have custom logic in the
destructor of the FFTRealTransform class.

Finally, I removed the effort private member, as there is was no point
in storing that information really.

It should be noted that we could have something similar for the plan
objects, but we're tackling one at a time I guess.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>