Defining CL_TARGET_OPENCL_VERSION macro
This macro was introduced recently in the new unified version of the
OpenCL headers as published by Khronos, which were previously kept
separately for each version (and therefore did not require the user to
point for a particular version). Not having it makes the header files
default to 2.2, but also to generate a warning message. Having it
doesn't hurt for previous versions of the OpenCL headers.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>