Warnings out for non-OpenMP builds
Newer compilers compilers are complaining (rightly so, via
-Wunused-private-field) about unused fields that are related to OpenMP.

In the case of the brute-force convolver classes, this field is passed
down to omp_2d_for, and therefore it's easier to use the UNUSED() at
that level. I personally haven't been able to reproduce these two
warnings (one for BruteForceConvolver and other for
AssosiativeBruteForceConvolver), but we should see them in some versions
of clang++.

In the case of FFTPlan, we previously used to store the omp_thread
number to support copy constructors (if I remember correctly), but since
moving to using unique_ptrs for storing the plans there is no need
anymore to store the omp_thread count anymore.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>