Moving headers into include/ directory
This was a change I wanted to do a long time ago, but I only got around
doing it now, and only because some pressure was put onto us. It turns
that in C++20 a new 'version' header was added to the set of standard
headers. This name clashes with our VERSION file in case-insensitive
filesystems like MacOS, and therefore when running new compilers like
clang8 8.0 in MacOS code fails to compile due to the new 'version'
standard header being eventually included internally by some of the
other standard headers.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>