Trying to fix MacOS Xcode 7.3 builds
Homebrew is not distributing binary bottles anymore for some (most?)
El Capitan packages, and thus we are being left with having to compile
them on our own, which takes longer than Travis would like. Let's thus
simply skip the installation of fftw (which pulls gcc and open-mpi) in
this flavour of Xcode, which will allow us to continue testing this
platform, though less intensively.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>