New stoui and ceil_div utility functions
These are meant to be used in a couple of places in the code. The former
is specially useful since we use unsigned int values in lots of places.
Our implementation of stoui assumes the unsigned long values actually
fit into unsigned ints, which is fine for our library at the moment.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>