Build libprofit statically in Windows
Even though I put some effort on turning libprofit into a proper shared
library in Windows, it's clear now to me that this was a wrong approach.
Providing the symbols' visibilities is not enough -- the entire API must
be exportable, and that doesn *not* include classes and types that are
defined outside of our control, like anything inside the std namespace.
In other words, even though I'm exporting symbols, the resulting library
is as good as a static library in terms of how reusable it is.

I'm changing the Windows target for the time being, because in
Linux/MacOS we are still relying on shared libraries to compile
pyprofit. We can tackle those later.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>