Using unique_ptr to handle fftw_plans
This is something I wanted to do for a while, but never got around doing
it until I saw it now being highlighted by sonar qube. The tricky thing
is that fftw doesn't specify the type pointed to by the fftw_plan type,
so a little bit of digging was necessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>