Removing unnecessary overloads
These are automatically generated by the compiler already so there is no
need for us to define them explicitly. In fact our explicit definitions
created more trouble than benefit, since we were defining the copy/move
ctors/assignments but we were not defining a destructor, and therefore
were violating the Rule of Five.

The only time where we want to maintain these explicit functions is in
the _2dcoordinate class, where we want to give some special meaning to
move-from objects so they become (0, 0) 2d coordinates. In this case we
don't remove the special functions, but actually add the destructor to
complete the Five.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>