New surface.revert() method
This takes what we were previously doing externally in the
AssociativeBruteForceConvolver class (reverting the kernel's image) and
doing it internally within the surface class.

The previous code was also unnecessarily creating a vector first,
reverting that, and then creating an Image by moving the vector. This
was unnecessary, as one could have used the begin/end methods of the
Image itself and pass them down to the std::reverse method, which is
what we're doing now.

Finally, a new unit test checks that this is correctly implemented.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>