Restructuring Convolver object inner workings
Instead of declaring a virtual convolve method on the Convolver class
and implementing it in all subclasses, I now implement that method in
the Convolver class and defer the implementation to the new virtual
convolve_impl method. That will allows us to put code common to all
convolution methods in a central place.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>