Adding support for convolving krn > image
This was not supported before; the FFTConvolver properly checked this,
but the brute force convolvers didn't and simply crashed. The OpenCL
convolvers worked only on certain occasions due to the internal padding
they already do.

This first implementation doesn't take into account the padding
potentially done internally by each convolution method, so there is
still room for improvement. Anyway, kernels bigger than the images they
are being convolved with are rare, so the situation is not tragic.

A new battery of tests checks that now we can perform convolution of
images that are smaller than kernels, and that the resulting image is
still the same size of the original.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>