Extract convolver creation into new method
When the model needs to use convolver, it uses the one given by the
user, if any, or is creates one internally. This process of ensuring a
convolver is present when needed was locally done when convolving the
image coming out of the profile evaluation, but when adding automatic
psf-driven padding to the image sizes we'll need to use the convolver
earlier during the process of the model evaluation; thus, it makes sense
to have the logic of ensuring a convolver in a reusable form.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>