Introducing the new PointPair and Box classes
The first is simply a pair of Points with no further constrains. The
latter derives from the former, and is used explicitly to define a box.
In this case the Point giving the upper bound is exclusive, and
therefore an empty Box can be defined by giving the same points as the
lower and upper bounds.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>