New surface::bounding_box method
This surface scans a surface, checks for values different from zero in
it, and returns the minimum box within that surface containing all
non-zero values. This will be useful when inspecting masks and deciding
whether they are small enough to avoid having to generate a model image
bigger than requested in order to preserve flux.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>