Adding fft convolution tests for sky/null profiles
These were not present, and although they are not *that* important they
still represent an example of a non-radial profile being convolved. This
is of significance to check whether non-radial profiles work as expected
in a number of situations involving convolution, like the extra
psf-induced padding we want to introduce.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>