Adding Mask upsampling functionality
This is very similar to Image's upsampling functionality, except that in
the case of a Mask we support a single "mode" consisting on copying cell
values to the corresponding cells (unlike Image upsampling, which also
supports flux distribution across the new pixels).

Since the basic algorithm is the same I'm extracting away the logic for
upsampling and reusing it for both Image and Mask upsampling.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>