New FFTRealTransformer resizing functionality
Until now the FFTRealTransformer class has been quite static in what it
can do, leaving no flexibility to change the data sizes over which it
operates. This commit changes things, and introduces a new resize()
method that, given a new size, makes the necessary adjustments on the
internal members of the class to process data of a given new size (but
the effort and number of threads remain the same). Thus, things that
need changing are the plan sizes (and therefore plans need to be
re-calculated), scratch buffer allocations and references sizes.

When the new requested size is the same as before, no reallocations and
calculations happen. This makes client code easier to write, as no
special care needs to be taken for these corner cases. Giving a size of
0 is not permitted.

Given this flexibility on changing the size, a new constructor has also
been added that creates a new transformer without a given size.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>