Resizing ability for the FFTConvolver
Like with the FFTRealTransformer, we are now giving a resizing ability
to the FFTConvolver class. This means that now users don't need to give
in advance the exact sizes that are going to be used by the convolver,
as it can at runtime adjust itself to newly given sizes for the source
and kernel images.

When resizing, both the scratch fft data buffers, and the extended
scratch Images have to be recalculated. In addition, if kernel reusing
is activated we need to correctly flag that a new forward calculation
needs to happen.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>