New Mask::expand_by method
This method is similar in spirit to a convolution, as the true values of
the mask are copied into adjacent cells based on the given extents. A
"simpler" implementation would have been to transform the mask to
numerical values (0, 1), convolve against a psf, and convert back to
boolean values using a "value != 0" condition. This implementation is a
bit simpler though, as there is no need to perform any sum, and we can
break earlier.

New unit tests check for this behavior. In particular we added a new
assert_masks method to check that two masks are equal, and check where
they differ. This new method required a small hack to avoid problems in
Apple's clang compiler, where the compiler refused to cast from an
internal std::__bit_iterator template class type to unsigned char *.
It's unclear to me why this happens, so I'm letting it go for the time
being.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>