Model::evaluate broken down into smaller functions
The complexity of Model::evaluate was already a bit high, and with the
introduction in the future of the flux capturing feature this will only
increase. This commit then breaks down the current evaluate function
into three: evaluate, which still contains the high-level logic of the
evaluation process; analyze_inputs, which does all the parameter
validation and common input-based calculations needed by the model
evaluation; and produce_image, which contains the core logic of
generating the profile images and convolving them.

I also took the opportunity to do some small code adjustments to make it
more readable. This includes improved spacing, and usage of standard
algorithms.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>