Correctly capturing flux when convolving
This is the final bit needed to implement correct flux capturing in
libprofit. Until now libprofit didn't capture flux outside of the
"drawing" region ("outside" of the model, or outside of the mask when a
mask is present) when performing convolution. This however is a feature
that needs to be taken into account if we want to correctly generate
convolved images.

To capture flux correctly we need to enlarge the "drawing" area in such
a way that when convolution gets done we can capture the flux outside of
the original model boundaries. When no mask is involved we enlarge the
model by the half of the PSF dimensions on each side, and let the
profiles know about the new origin. Additionally, when a mask is
involved, we need to enlarge *its* coverage. Finally, if a mask is
sufficiently small we can avoid enlarging the model altogether, but we
still need to convolve the mask.

This image enlargement is something new we have added to libprofit, and
it's probably what the "no crop" feature on the R side initially was
meant to be for. However, in libprofit so far we have had the "no crop"
feature to mean exactly "don't crop enlargements coming from the
convolution step", because there was nothing else to crop out. This
behaviour will be preserved for the time being, but probably a change in
semantics will be required in the future to make the code simpler.

New unit tests have been added to make sure this feature is working
across our different convolvers, for different mask sizes and
placements, and with/without finesampling enabled. The check we perform
is simple: with a given psf we generate a model image (A); we then use
masks of different sizes and at different placements and generate a
model image (B) for each of them. For all these mask combinations the
image pixels *within the mask's region* should be the exact same for A
and B.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>