Correcting Image/Mask value constructors
The value + Dimensions constructor of the Image and Mask classes was
buggy because the dimensions argument was being moved (which causes it
to be zero'd out) potentially before being used to construct the vector
objects using the dimension's x and y elements. These two operations are
performed in a non-deterministic way because they both happen as part of
the argument evaluation of the parent surface constructor invocation. It
wasn't until until C++17 (I think) that this order was guaranteed to be
from left to right.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>