Check mask dimensions are as expected
This has always been an implicit requirement for the rest of the code to
work correctly and until now we had never required it explicitly.
Luckily no test actually breaks with this change, which means we have
been doing things correctly so far.

A new unit test checks that these dimensions are indeed checked, and
that the appropriate error is thrown whenever a mask with wrong
dimensions is used.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>