Refactoring mask calculation code for reuse
The new code we introduced to automatically expand mask and its coverage
depending on the model's inputs is being broken down into three
different re-usable pieces of code: a) calculate the parts of the input
analysis that are related to this expansion process (the new
analyze_expansion_requirements function), b) check if the mask needs
any type of modification (needs_adjustment) and c) actually perform the
necessary adjustments to the mask.

The reason to provide this broken down behavior is that we want to offer
a new, static method that, given a Mask, a PSF, a model dimension and a
finesampling factor, it will return a modified version of the Mask that
is fit for usage in those circumstances, so that a Model can consume it
as-is without internally modifying it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>