Enable usage of pre-computed Masks
While the previous commit allowed users to pre-compute the Mask that a
Model would use internally for its evaluation, this commit allows them
to inform the Model that it shouldn't compute a new Mask internally, and
that the one given by the user should be used as-is instead. When this
happens we still check that the mask dimensions are what we expect them
to be, which is the only check we can safely perform.

This functionality means we can reuse an "offline" computed mask acros
model evaluations that use the same PSF, Mask, finesampling factor and
model sizes (e.g., during a high-level profile fitting optimization),
saving us from having to calculate the internal mask every time.

A unit test checks that using the offline computed mask or the
internally computed one doesn't make a difference, as the final result
is the same.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>