Avoid copying OpenCL binaries
This is not exactly a problem per-se, and the binaries are probably not
too big, but I got my attention drawn into this after reading a valgrind
report complaining about the subsequent output.write spuriously pointing
to uninitialized bytes that had been initialized during a "new" call in
the KernelCache::to_cache() function. Again, even with copying things
should have worked as expected, but it's better to avoid the copy
altogether.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>