Improving memory usage in Model::produce_image
Until now the Model class had been accumulating the individual images
coming out from each profile (i.e., putting them into a vector). After
this they were reduced (summed up), and convolved as required. This was
to make the code simpler to write in the first place, but has the big
consequence that it imposes the need for memory that is otherwise
unnecessary.

This commit replaces the previous strategy with a much simpler one: have
two "final" image buffers (one for images that require convolution,
another for those that don't), plus a scratch space. When iterating over
the profiles, the scratch space is filled with zeroes and given to the
profile. After profile evaluation, the scratch is added into the
corresponding buffer. Finally, the buffer containing images to be
convolved is effectively convolved, and then added into the other buffer
to make the final image.

This new strategy requires a fixed amount of memory, and thus it can
handle many more profiles than we could previously do. It also actually
has an impact on the runtime as less memory allocations and relinquishes
are necessary. The code is also structured in such a way that if no
convolution is ever required then the buffer that would contain the
images requiring convolution is not allocated at all, and that final
addition of images also doesn't happen.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>